Restructure repo for web/node builds
