+++ source/base/shared.ts
@@ -1,3 +1,4 @@
 import { EncryptionType } from "./constructs";
 
 export const ALGO_DEFAULT = EncryptionType.CBC;
+export const DERIVED_KEY_ITERATIONS = 250000;
