+++ source/base/timing.ts
