+++ test/unit/encryption.spec.js
@@ -5,8 +5,9 @@ const {
     encryptGCM,
     generateIV,
     generateSalt
-} = require("../../dist/encryption.js");
-const { deriveFromPassword, pbkdf2 } = require("../../dist/derivation.js");
+} = require("../../dist/node/encryption.js");
+const { deriveFromPassword } = require("../../dist/base/derivation.js");
+const { pbkdf2 } = require("../../dist/node/derivation.js");
 
 const ENCRYPTED_SAMPLE = "at5427PQdplGgZgcmIjy/Fv0xZaiKO+bzmY7NsnYj90=";
 const ENCRYPTED_SAMPLE_RAW = "iocane secret text";
